[BugFix] Fix 0 destination node error in HGT encoder (#1337)
*Issue #, if available:*

*Description of changes:*
This PR fix the issue of mismatch between number of nodes and number of
features when the number of destination nodes is 0.
This PR also added a new test for this corner case for all three
built-in hetero encoder.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-1-12.us-west-2.compute.internal>