[Feature] Add edge features support for homogenous GNN: GAT (#1349)
*Issue #, if available:*

*Description of changes:*
This PR adds the usage of edge_feat in the GAT model for homogeneous
graph in gsf.py and gat_encoder.py. The corresponding unit-test is added
in test_nn_encoder.py and test_nn_model.py.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-81-245.ec2.internal>