[Enhancement] Support Learnable Embeddings for Real-time inference (#1363)
*Issue #, if available:*

*Description of changes:*
This PR adds support of a special node feature, `gs_embedding`, in the
payload JSON for real-time inference case.

This works when the deployed GraphStorm model was trained with learnable
embeddings in the two cases below:
1. some node types are featureless, hence GraphStorm automatically
assigned learnable embeddings on them while training; and,
2. user set the `--use-node-embedding` configuration to be true while
training.

After trained, these learnable embeddings will be saved under different
folders named with node type names. In offline inference, GraphStorm
will load them as part of model artifacts. Because size of these
learnable embeddings are proportional to the number of nodes, it could
be hundreds of MBs or a few GBs. Reload these learnable embeddings on
SageMaker endpoints could cause potential risks, like OOM and OOD.

Major changes:
- create two new input layer encoders,
`GSPureLearnableInputLayer4GraphFromMetaData`, and
`GSNodeEncoderInputLayer4GraphFromMetadata` to handle models that cannot
load learnable embeddings at SageMaker endpoints.
- modify the `prepare_batch_input()` method to extract the
`gs_embedding` feature and prepare it for model's forward function.
- modify the `gsf.py` model building to replace node input layers when
the initial graphs are coming from metadata.
- modify the `process_json_payload_nodes()` function to allow payload
graph to contains features that was not used in model training, which is
a common case because model training uses `--node-feat-name` to control
which features to be used.

Tests:
- in local SageMaker environment.
- in Amazon SageMaker environment.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-60-50.us-west-2.compute.internal>
Co-authored-by: jalencato <jalencato23pistons@gmail.com>